+++ src/polygon-rasterizer.h
@@ -39,7 +39,7 @@ typedef struct {
 
 row_crossings_t *get_row_crossings(mpoly_t *mpoly, int min_y, int num_rows);
 void free_row_crossings(row_crossings_t *rc, int num_rows);
-void mask_from_mpoly(mpoly_t *mpoly, int w, int h, const char *fn);
+void mask_from_mpoly(mpoly_t *mpoly, size_t w, size_t h, const char *fn);
 void crossings_intersection(row_crossings_t *out, row_crossings_t *in1, row_crossings_t *in2);
 
 #endif // ifndef POLYGON_RASTERIZER_H
